Basic structure. 3.0.2 1. ReviewServiceImpl  1-1. implement leaveReview  1-2. implement delete  1-3. implement getReviewByDocumentId  1-4. implement getRatingForDocument 2. Refactoring Test Suite and cases

+++ src/backend/java/com/siwoo/document_application/service/AbstractApplicationService.java
@@ -1,13 +1,14 @@
 package com.siwoo.document_application.service;
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 public abstract class AbstractApplicationService<T,ID>  {
 
     /*generic repository*/
@@ -38,8 +39,11 @@ public abstract class AbstractApplicationService<T,ID>  {
 
     /*return true if domain is deleted otherwise false*/
     boolean delete(ID id){
+        if(id == null) return false;
+        if(!get(id).isPresent()) return false;
+
         getRepository().deleteById(id);
-        return !getRepository().existsById(id);
+        return true;
     }
